% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close.R
\name{close}
\alias{close}
\title{close}
\usage{
close(kind = "pdf", return.text = FALSE)
}
\arguments{
\item{kind}{Vector, among c("windows", "quartz", "x11", "X11", "pdf", "bmp", "png", "tiff"), indicating the kind of graphic windows (devices) to close. BEWARE: either "windows", "quartz", "x11" or "X11" means that all the X11 GUI graphics devices will be closed, whatever the OS used.}

\item{return.text}{Single logical value. Print text regarding the kind parameter and the devices that were finally closed?}
}
\value{
Text regarding the kind parameter and the devices that were finally closed.
}
\description{
Close only specific graphic windows (devices).
}
\details{
REQUIRED PACKAGES

cuteDev


REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

arg_check()
}
\examples{
windows() ; 
windows() ; 
pdf() ; 
dev.list() ; 
close(kind = c("pdf", "x11"), return.text = TRUE) ; 
dev.list()
}
