\name{cuteGraph-package}
\alias{cuteGraph-package}
\alias{cuteGraph}
\docType{package}
\title{
\packageTitle{cuteGraph}
}
\description{
\packageDescription{cuteGraph}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cuteGraph}
\packageIndices{cuteGraph}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cuteGraph}

Maintainer: \packageMaintainer{cuteGraph}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
