% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal_functions.R
\name{.cuteDev_package_check}
\alias{.cuteDev_package_check}
\title{.cuteDev_package_check}
\usage{
.cuteDev_package_check(req.package = "cuteDev", lib.path = NULL)
}
\arguments{
\item{req.package}{Character vector of package names to import.}

\item{lib.path}{Optional character vector specifying the absolute pathways of the directories containing some of the listed packages in the req.package argument, if not in the default directories. Ignored if NULL.}
}
\value{
Nothing.
}
\description{
Check if the cuteDev packages is present in the computer.
}
\keyword{internal}
