% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open2.R
\name{open2}
\alias{open2}
\title{open2}
\usage{
open2(
  pdf = TRUE,
  pdf_path = "working.dir",
  pdf_name = "graph",
  width = 7,
  height = 7,
  paper = "special",
  pdf_overwrite = FALSE,
  rescale = "fixed",
  remove_read_only = TRUE,
  return_output = TRUE,
  safer_check = TRUE,
  lib_path = NULL,
  error_text = ""
)
}
\arguments{
\item{pdf}{Single logical value. Use pdf display? If FALSE, a GUI is opened.}

\item{pdf_path}{Single character string. Where the pdf is saved (do not terminate by / or \\). Write "working.dir" if working directory is required (default). Ignored if pdf == FALSE.}

\item{pdf_name}{Single character string. Name of the pdf file containing the graphs (the .pdf extension is added by the function, if not detected in the name end). Ignored if pdf == FALSE.}

\item{width}{Single positive numeric value indicating the width of the window (in inches).}

\item{height}{Single positive numeric value indicating the height of the window (in inches).}

\item{paper}{Single character string. Paper argument of the pdf function (paper format). Only used for pdf(). Either "a4", "letter", "legal", "us", "executive", "a4r", "USr" or "special". If "special", means that the paper dimension will be width and height. With another paper format, if width or height is over the size of the paper, width or height will be modified such that the plot is adjusted to the paper dimension (see $dim in the returned list below to see the modified dimensions). Ignored if pdf == FALSE.}

\item{pdf_overwrite}{Single logical value. Existing pdf can be overwritten? . Ignored if pdf == FALSE.}

\item{rescale}{Kind of GUI. Either "R", "fit", or "fixed". Ignored on Mac and Linux OS. See ?windows for details.}

\item{remove_read_only}{Single logical value. Remove the read only (R.O.) graphical parameters? If TRUE, the graphical parameters are returned without the R.O. parameters. The returned $ini.par list can be used to set the par() of a new graphical device. If FALSE, graphical parameters are returned with the R.O. parameters, which provides information like text dimension (see ?par() ). The returned $ini.par list can be used to set the par() of a new graphical device, but generate a warning message. Ignored if return_output == FALSE.}

\item{return_output}{Single logical value. Return output ? If TRUE the output list is displayed.}

\item{safer_check}{Single logical value. Perform the "safer" checks? If \code{TRUE}, checkings are performed before main code running (see the \href{https://github.com/safer-r}{safer-r project}): 1) correct \code{lib_path} argument value 2) required functions and related packages effectively present in local R libraries and 3) R classical operators (like \code{"<-"}) not overwritten by another package because of the R scope. Warning: must be set to \code{FALSE} if this function is used inside another "safer" function to avoid pointless multiple checkings.}

\item{lib_path}{Vector of characters specifying the absolute pathways of the directories containing the required packages for the function, if not in the default directories. Useful when R packages are not installed in the default directories because of lack of admin rights. More precisely, \code{lib_path} is passed through the \code{new} argument of \code{.libPaths()} so that the new library paths are \code{c(lib_path, .libPaths())}. Warning: \code{.libPaths()} is restored to the initial paths, after function execution. Ignored if \code{NULL} (default) or if the \code{safer_check} argument is \code{FALSE}: only the pathways specified by the current \code{.libPaths()} are used for package calling.}

\item{error_text}{Single character string used to add information in error messages returned by the function, notably if the function is inside other functions, which is practical for debugging. Example: \code{error_text = " INSIDE <PACKAGE_1>::<FUNCTION_1> INSIDE <PACKAGE_2>::<FUNCTION_2>."}. If \code{NULL}, converted into \code{""}.}
}
\value{
A list containing:

- $pdf.loc: path of the pdf created.

- $ini.par: initial par() parameters.

- $zone.ini: initial window spliting.

- $dim: dimension of the graphical device (in inches).
}
\description{
Open a pdf or screen (GUI) graphic window and return initial graphic parameters.

This order can be used:

width()

open2()

prior_plot() # not for ggplot2

plot() or any other plotting

post_plot() if prior_plot() has been used # not for ggplot2

close2()
}
\details{
WARNINGS

On Linux, use pdf = TRUE, if (GUI) graphic window is not always available, meaning that X is not installed (clusters for instance). Use X11() in R to test if available.
}
\examples{
\dontrun{
# Screen devices should not be used in examples
open2(pdf = FALSE, pdf_path = ".", pdf_name = "graph", width = 7, height = 7, paper = "special", pdf_overwrite = FALSE, return_output = TRUE)
}
}
\author{
Gael Millot <gael.millot@pasteur.fr>

Yushi Han <yushi.han2000@gmail.com>

Haiding Wang <wanghaiding442@gmail.com>
}
