% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.R
\name{scale}
\alias{scale}
\title{scale}
\usage{
scale(n, lim, kind = "approx", lib.path = NULL)
}
\arguments{
\item{n}{Single positive and non null integer value indicating the desired number of main ticks on the axis.}

\item{lim}{Vector of 2 numbers indicating the limit range of the axis. Order of the 2 values matters (for inverted axis). Can be log transformed values.}

\item{kind}{Single character string. Either "approx" (approximative), "strict" (strict) or "strict.cl" (strict clean). If "approx", use the scales::trans_breaks() function to provide an easy to read scale of approximately n ticks spanning the range of the lim argument. If "strict", cut the range of the lim argument into n + 1 equidistant part and return the n numbers at each boundary. This often generates numbers uneasy to read. If "strict.cl", provide an easy to read scale of exactly n ticks, but sometimes not completely spanning the range of the lim argument.}

\item{lib.path}{Character vector specifying the absolute pathways of the directories containing the required packages if not in the default directories. Ignored if NULL.}
}
\value{
A vector of numbers.
}
\description{
Attempt to select nice scale numbers when setting n ticks on a lim axis range.
}
\details{
REQUIRED PACKAGES

cuteDev

cuteTool

if kind = "approx":

ggplot2

scales


REQUIRED FUNCTIONS FROM CUTE_LITTLE_R_FUNCTION

arg_check()

round()
}
\examples{
# approximate number of main ticks

ymin = 2 ; 
ymax = 3.101 ; 
n = 5 ; 
scale <- scale(n = n, lim = c(ymin, ymax), kind = "approx") ; 
scale ; 
par(yaxt = "n", yaxs = "i", las = 1) ; 
plot(ymin:ymax, ymin:ymax, xlim = range(scale, ymin, ymax)[order(c(ymin, ymax))], ylim = range(scale, ymin, ymax)[order(c(ymin, ymax))], xlab = "DEFAULT SCALE", ylab = "NEW SCALE") ;
par(yaxt = "s") ; 
axis(side = 2, at = scale)


# strict number of main ticks

ymin = 2 ; 
ymax = 3.101 ; 
n = 5 ; 
scale <- scale(n = n, lim = c(ymin, ymax), kind = "strict") ; 
scale ; 
par(yaxt = "n", yaxs = "i", las = 1) ; 
plot(ymin:ymax, ymin:ymax, xlim = range(scale, ymin, ymax)[order(c(ymin, ymax))], ylim = range(scale, ymin, ymax)[order(c(ymin, ymax))], xlab = "DEFAULT SCALE", ylab = "NEW SCALE") ; 
par(yaxt = "s") ; 
axis(side = 2, at = scale)


# strict "clean" number of main ticks

ymin = 2 ; 
ymax = 3.101 ; 
n = 5 ; 
scale <- scale(n = n, lim = c(ymin, ymax), kind = "strict.cl") ; 
scale ; 
par(yaxt = "n", yaxs = "i", las = 1) ; 
plot(ymin:ymax, ymin:ymax, xlim = range(scale, ymin, ymax)[order(c(ymin, ymax))], ylim = range(scale, ymin, ymax)[order(c(ymin, ymax))], xlab = "DEFAULT SCALE", ylab = "NEW SCALE") ; 
par(yaxt = "s") ; 
axis(side = 2, at = scale)


# approximate number of main ticks, scale inversion

ymin = 3.101 ; 
ymax = 2 ; 
n = 5 ; 
scale <- scale(n = n, lim = c(ymin, ymax), kind = "approx") ; 
scale ; 
par(yaxt = "n", yaxs = "i", las = 1) ; 
plot(ymin:ymax, ymin:ymax, xlim = range(scale, ymin, ymax)[order(c(ymin, ymax))], ylim = range(scale, ymin, ymax)[order(c(ymin, ymax))], xlab = "DEFAULT SCALE", ylab = "NEW SCALE") ; 
par(yaxt = "s") ; 
axis(side = 2, at = scale)
}
