% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/close2.R
\name{close2}
\alias{close2}
\title{close2}
\usage{
close2(kind = "pdf", return.text = FALSE)
}
\arguments{
\item{kind}{Vector, among c("windows", "quartz", "x11", "X11", "pdf", "bmp", "png", "tiff"), indicating the kind of graphic windows (devices) to close. BEWARE: either "windows", "quartz", "x11" or "X11" means that all the X11 GUI graphics devices will be closed, whatever the OS used.}

\item{return.text}{Single logical value. Print text regarding the kind parameter and the devices that were finally closed?}
}
\value{
Text regarding the kind parameter and the devices that were finally closed.
}
\description{
Close only specific graphic windows (devices).
}
\examples{

# Screen devices (windows(), quartz() and x11()) should not be used in examples
postscript(NULL) # open a postscript graphic device
pdf(NULL) # open a pdf graphic device
postscript(NULL) # open a postscript graphic device
pdf(NULL) # open a pdf graphic device
grDevices::dev.list() ; 
close2(kind = c("pdf"), return.text = TRUE) ; 
grDevices::dev.list() # only remains the postscript devices
grDevices::graphics.off()
}
