% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empty_graph.R
\name{empty_graph}
\alias{empty_graph}
\title{empty_graph}
\usage{
empty_graph(text = NULL, text.size = 1, title = NULL, title.size = 1.5)
}
\arguments{
\item{text}{Single character string of the message to display.}

\item{text.size}{Single numeric value of the text size.}

\item{title}{Single character string of the graph title.}

\item{title.size}{Single numeric value of the title size (in points).}
}
\value{
An empty plot.
}
\description{
Display an empty plot with a text in the middle of the window (for instance to specify that no plot can be drawn).
}
\examples{
# simple example

empty_graph(text = "NO GRAPH")


# white page

empty_graph() # white page


# all the arguments

empty_graph(text = "NO GRAPH", text.size = 2, title = "GRAPH1", title.size = 1)

}
