% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inter_ticks.R
\name{inter_ticks}
\alias{inter_ticks}
\title{inter_ticks}
\usage{
inter_ticks(
  lim,
  log = "log10",
  breaks = NULL,
  n = NULL,
  warn.print = TRUE,
  safer_check = TRUE
)
}
\arguments{
\item{lim}{Vector of 2 numbers indicating the limit range of the axis. Order of the 2 values matters (for inverted axis). If log argument is "log2" or "log10", values in lim must be already log transformed. Thus, negative or zero values are allowed.}

\item{log}{Single character string. Either "log2" (values in the lim argument are log2 transformed) or "log10" (values in the lim argument are log10 transformed), or "no".}

\item{breaks}{Mandatory vector of numbers indicating the main ticks values/positions when log argument is "no". Ignored when log argument is "log2" or "log10".}

\item{n}{Single numeric value indicating the number of secondary ticks between each main tick when log argument is "no". Ignored when log argument is "log2" or "log10".}

\item{warn.print}{Single logical value. Print potential warning messages at the end of the execution? If FALSE, warning messages are never printed, but can still be recovered in the returned list.}

\item{safer_check}{Single logical value. Perform some "safer" checks (see https://github.com/safer-r)? If TRUE, checkings are performed before main code running: 1) R classical operators (like "<-") not overwritten by another package because of the R scope and 2) required functions and related packages effectively present in local R lybraries. Set to FALSE if this fonction is used inside another "safer" function to avoid pointless multiple checkings.}
}
\value{
A list containing :

- $log: value of the log argument used.

- $coordinates: the coordinates of the secondary ticks on the axis, between the lim values.

- $values: the corresponding values associated to each coordinate (with log scale, 2^$values or 10^$values is equivalent to the labels of the axis).

- $warn: the potential warning messages. Use cat() for proper display. NULL if no warning.
}
\description{
Define coordinates and values of secondary ticks.
}
\examples{
# no log scale

inter_ticks(lim = c(-4,4), log = "no", breaks = c(-2, 0, 2), n = 3)
inter_ticks(lim = c(10, 0), log = "no", breaks = c(10, 8, 6, 4, 2, 0), n = 4)


# log2

inter_ticks(lim = c(-4,4), log = "log2")


# log10

inter_ticks(lim = c(-2,3), log = "log10")
}
